Check if query exists in endpoint

+++ src/endpoint/Item.php
@@ -14,29 +14,29 @@ class Item extends Endpoint {
    * @throws DwapiException
    */
   public function get() {
+    if ($this->query) {
+      $this->query->property = $this->request->getParameters("query", "property", true, true, false);
+      $this->query->relation = $this->request->getParameters("query", "relation", true, true, false);
+      $this->query->hash = $this->request->getParameters("path", "hash");
+      if (!is_null($this->query->hash)) {
+        $this->query->id = Helper::getIdFromHash($this->query->hash);
 
-    $this->query->property = $this->request->getParameters("query", "property", true, true, false);
-    $this->query->relation = $this->request->getParameters("query", "relation", true, true, false);
-    $this->query->hash = $this->request->getParameters("path", "hash");
-    if (!is_null($this->query->hash)) {
-      $this->query->id = Helper::getIdFromHash($this->query->hash);
+        $this->query->single_read();
+      } else {
+        $this->query->filter = $this->request->getParameters("query", "filter", true, true, false);
+        $this->query->paging = $this->request->getParameters("query", "paging", false, false, false);
+        $this->query->sort = $this->request->getParameters("query", "sort", true, true, false);
 
-      $this->query->single_read();
+        $this->query->read();
+      }
+
+      $this->result = $this->query->getResult();
+      if ($this->current_token && $this->current_token->token_type == "jwt") {
+        $this->result["extended_token"] = $this->current_token->extend_token();
+      }
+
+      $this->debug = $this->query->getDebug();
     }
-    else {
-      $this->query->filter = $this->request->getParameters("query", "filter", true, true, false);
-      $this->query->paging = $this->request->getParameters("query", "paging", false, false, false);
-      $this->query->sort = $this->request->getParameters("query", "sort", true, true, false);
-
-      $this->query->read();
-    }
-
-    $this->result = $this->query->getResult();
-    if ($this->current_token && $this->current_token->token_type == "jwt") {
-      $this->result["extended_token"] = $this->current_token->extend_token();
-    }
-
-    $this->debug = $this->query->getDebug();
   }
 
   /**
@@ -44,28 +44,28 @@ class Item extends Endpoint {
    * @throws DwapiException
    */
   public function put() {
+    if ($this->query) {
+      $this->query->hash = $this->request->getParameters("path", "hash");
+      $this->query->values = $this->request->getParameters("formData", NULL, true, false, true);
 
-    $this->query->hash = $this->request->getParameters("path", "hash");
-    $this->query->values = $this->request->getParameters("formData", NULL, true, false, true);
+      $this->request->processFiles($this->query->values);
 
-    $this->request->processFiles($this->query->values);
+      if (!is_null($this->query->hash)) {
+        $this->query->id = Helper::getIdFromHash($this->query->hash);
+        $this->query->single_update();
 
-    if (!is_null($this->query->hash)) {
-      $this->query->id = Helper::getIdFromHash($this->query->hash);
-      $this->query->single_update();
+      } else {
+        $this->query->filter = $this->request->getParameters("query", "filter", true, true, true);
+        $this->query->update();
+      }
 
+      $this->result = $this->query->getResult();
+      if ($this->current_token && $this->current_token->token_type == "jwt") {
+        $this->result["extended_token"] = $this->current_token->extend_token();
+      }
+
+      $this->debug = $this->query->getDebug();
     }
-    else {
-      $this->query->filter = $this->request->getParameters("query", "filter", true, true, true);
-      $this->query->update();
-    }
-
-    $this->result = $this->query->getResult();
-    if ($this->current_token && $this->current_token->token_type == "jwt") {
-      $this->result["extended_token"] = $this->current_token->extend_token();
-    }
-
-    $this->debug = $this->query->getDebug();
   }
 
   /**
@@ -74,23 +74,23 @@ class Item extends Endpoint {
    */
   public function post()
   {
-    $this->query->values = $this->request->getParameters("formData", NULL, true, false, true);
-    $this->request->processFiles($this->query->values);
+    if ($this->query) {
+      $this->query->values = $this->request->getParameters("formData", NULL, true, false, true);
+      $this->request->processFiles($this->query->values);
 
-    if ($this->query->create()) {
-      $this->http_response_code = 201;
-      $this->result = $this->query->getResult();
-      if ($this->current_token && $this->current_token->token_type == "jwt") {
-        $this->result["extended_token"] = $this->current_token->extend_token();
+      if ($this->query->create()) {
+        $this->http_response_code = 201;
+        $this->result = $this->query->getResult();
+        if ($this->current_token && $this->current_token->token_type == "jwt") {
+          $this->result["extended_token"] = $this->current_token->extend_token();
+        }
+
+        $this->debug = $this->query->getDebug();
+        return;
+      } else {
+        $this->result = array("id" => NULL);
       }
-
-      $this->debug = $this->query->getDebug();
-      return;
     }
-    else {
-      $this->result = array("id" => NULL);
-    }
-
   }
 
   /**
@@ -98,23 +98,24 @@ class Item extends Endpoint {
    * @throws DwapiException
    */
   public function delete() {
-    $this->query->hash = $this->request->getParameters("path", "hash");
-    if (!is_null($this->query->hash)) {
-      $this->query->id = Helper::getIdFromHash($this->query->hash);
+    if ($this->query) {
+      $this->query->hash = $this->request->getParameters("path", "hash");
+      if (!is_null($this->query->hash)) {
+        $this->query->id = Helper::getIdFromHash($this->query->hash);
 
-      $this->query->single_delete();
+        $this->query->single_delete();
 
+      } else {
+        $this->query->filter = $this->request->getParameters("formData", "filter", true, false, true);
+        $this->query->delete();
+      }
+
+      $this->result = $this->query->getResult();
+      if ($this->current_token && $this->current_token->token_type == "jwt") {
+        $this->result["extended_token"] = $this->current_token->extend_token();
+      }
+
+      $this->debug = $this->query->getDebug();
     }
-    else {
-      $this->query->filter = $this->request->getParameters("formData", "filter", true, false, true);
-      $this->query->delete();
-    }
-
-    $this->result = $this->query->getResult();
-    if ($this->current_token && $this->current_token->token_type == "jwt") {
-      $this->result["extended_token"] = $this->current_token->extend_token();
-    }
-
-    $this->debug = $this->query->getDebug();
   }
 }
